Review the interface model
Basically, the synth parameters should come from an internal or external gear via midi.
Create a specific block to connect the knobs to constants representing the "external world", add a conversion block that scales the input parameters according to the midi control value and the internal requirements.
Inject the OSC2 saturation parameter and HPF cutoff frequency to start.